% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{assign_names}
\alias{assign_names}
\title{Assign optional name in global environment}
\usage{
assign_names(df, name)
}
\arguments{
\item{df}{A dataframe containing the stock variable(s) to be assigned a name in the global environment.}

\item{name}{The name for the output object in the global environment.}
}
\description{
Assign optional name in global environment
}
