% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{calc_stock}
\alias{calc_stock}
\title{Calculate stock}
\usage{
calc_stock(df, var, val)
}
\arguments{
\item{df}{A dataframe containing the stock variable(s) to be calculated.}

\item{var}{A character vector of variable names for which the stock measure will be created.}

\item{val}{The weights to apply to the cumulative sum of past values.}
}
\description{
Calculate stock
}
