% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{check_input}
\alias{check_input}
\title{Function to check all inputs to get_stock}
\usage{
check_input(var, val, fill, add, name)
}
\arguments{
\item{var}{The V-Dem variables you wish to calculate a stock measure from.}

\item{val}{The weights to apply to the cumulative sum of past values.}

\item{fill}{The number of years to fill forward missing values.}

\item{add}{Any additional V-Dem variables you wish added to the output dataset.}

\item{name}{An optional name for the output.}
}
\description{
Function to check all inputs to get_stock
}
