% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{create_antecedence}
\alias{create_antecedence}
\title{Create measure based on historical antecedence}
\usage{
create_antecedence(df, var, stock_tt)
}
\arguments{
\item{df}{A dataframe to be merged with the stock translation table.}

\item{var}{A character vector of variable names for which the stock measure will be created.}

\item{stock_tt}{The stock translation table, which contains the country-year combinations which will be changed to create stock_id.}
}
\description{
Create measure based on historical antecedence
}
