% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{create_stock_id}
\alias{create_stock_id}
\title{Creates stock_id according to the stock translation table}
\usage{
create_stock_id(df, stock_tt)
}
\arguments{
\item{df}{A dataframe to be merged with the stock translation table.}

\item{stock_tt}{The stock translation table, which contains the country-year combinations which will be changed to create stock_id.}
}
\description{
Creates stock_id according to the stock translation table
}
