% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{fill_data}
\alias{fill_data}
\title{Normalize and fill forward years}
\usage{
fill_data(df, var, fill)
}
\arguments{
\item{df}{A dataframe containing the stock variable(s) to be normalized and filled forward.}

\item{var}{A character vector of variable names for which the stock measure will be created.}

\item{fill}{The number of years to fill forward missing values.}
}
\description{
Normalize and fill forward years
}
