% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{subset_data}
\alias{subset_data}
\title{Subset variables and remove historical states}
\usage{
subset_data(df, var, add = NULL)
}
\arguments{
\item{df}{A dataframe to be subsetted.}

\item{var}{A character vector of variable names to be subsetted from the V-Dem dataset. Names must match variables names in the V-Dem dataset.}

\item{add}{Any additional V-Dem variables you wish added to the output dataset.}
}
\description{
Subset variables and remove historical states
}
